[Bugfix] Fix SQL issues and other minor bugs

+++ api/src/main/java/at/accounting_otter/entity/Debit.java
@@ -65,7 +65,7 @@ public class Debit {
         this.debtor = debtor;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
